src/web/requirements.txt
wheel
Flask
Flask-SQLAlchemy
flask_oauthlib
psycopg2-binary
